remove internal header from public
